+++ changelog.md
@@ -1,6 +1,7 @@
 0.1.21
 
 * Path graphics update.
+* Default to empt barrel icon for stickers. Can't handle tints yet.
 
 0.1.20